Add max_version parameter to SimpleGlobal
This is needed because new versions of protocols can change in a
backwards-incompatible manner, and clients unaware of the changes can
crash or behave incorrectly.

Note that Interface::VERSION on its own is not enough for this purpose,
because a protocol can make a backwards-incompatible change affecting
only the specified behaviour of the client without changing the interface
description.

Fixes https://github.com/Smithay/client-toolkit/issues/186