SlotPool: free buffers more aggressively
Currently, a buffer remains allocated in the pool as long as the server
has not run wl_display::delete_id on the buffer.  This is more
conservative than necessary, and sometimes prevents reusing the backing
memory of buffers that are known to not be in use.

This manually tracks the actual use of the slot and frees the memory
when it is no longer reachable by anyone.