RegistryState API rework
RegistryState::new constructs its own WlRegistry to reduce boilerplate
(it now requires a sync callback in addition to a WlRegistry).

RegistryState now handles the first display sync.  Examples have been
corrected to not assume a socket read provides a full round-trip - while
this may be true locally on Linux, it's not guaranteed and is likely to
be false if using waypipe.

The GlobalProxy enum handles the state of bound capability globals.

Fixes #186 by requiring callers to specify a maximum and checking that
against the maximum from the wayland XML.