Simplify PointerHandler
This exposes all the events in a pointer frame in one user callback
instead of making them distinct function calls.  It also provides
tracking of the active surface and pointer position for all events,
which is useful for button press/release.

Axis (scroll) events now correctly report both discrete and absolute
distances, instead of just the first of the two that was reported.