outputs: Add `logical_position` and `logical_size` properties
The documentation for `wl_output::mode` and `wl_output::geometry`
specifically states to use the information from these `xdg_output`
events for "logical" size and position within the "global compositor
space". Which may be impacted by scaling and transform. Given this, the
properties are not redundant and should be handled.

Testing under Sway, and looking at the implementation of wlroots, `x` and
`y` in the `geometry` event are always given as `0`, but the
`logical_position` has useful values.