Make `RegistryState` depend on use of `registry_queue_init`
This prevents the need to create two registry objects in examples like
`simple_window`, and eliminates the need to manually implement `Dispatch`.

This would perhaps make it harder if an application wants to handle
custom globals, instead of just putting that in the `Dispatch`
implementation, but if the `RegistryState` helper doesn't already
provide an ergonomic way to do that, maybe the way the
`ProvidesRegistryState` trait and `registry_handlers!` macro work isn't
ideal?

Leaving as a draft because I think it might be good to remove `ready()`
and make things like `OutputState::new()` also take a `&GlobalList`. But
does this seem like a generally good approach?