Remove `.ready()` from `RegistryState`; update `OutputState`/`SeatState`
This is inapplicable if `RegistryState::new()` is only called once the
globals are already "ready". This also makes
`RegistryState::from_registry` unusable, so it's removed (is that
important for anything?).

In `SeatState`, `new_seat()` is no longer called initially, and only for
seats added later. This doesn't impact any of the examples since they
don't do anything in that function. This is similar to how globals work.
Outputs are different since they aren't added until `wl_output::done`.

Since `OutputState::new` and `SeatState::new` don't have access to the
`RegistryState`, they can't call `RegistryState::{bind_all,bind_one}`.
So those are moved to private standalone functions. Is there a better
soliution here? It could just take an `&OutputState`, but that can be a
bit awkward.