keyboard: fix modifiers on re-issue and repeat
The modifiers state wasn't re-broadcasted when nothing got changed in
the effective state, which was incorrect for multi-window clients,
since it may spawn new window with depressed modifiers resulting
in new windows not seeing modifiers update event.

This commit also fixes the issue when key repeating stops due to
pressing SHIFT, while it should just alter the key and 'keep' the
repeat going.