window: use NonZeroU32 for WindowConfigure::new_size
The protocol states:

>If the width or height arguments are zero, it means the client should
>decide its own window dimension.

Meaning that either could be zero, though it's not clear whether
all dimensions should be picked, or just the one with zero.

It was also observed in the wild that some compositors could send
negative values for the configure, in which we should protect and
pick our own sizes to not cause overflow on casting.