Use map_copy_read_only() to map the keymap file
map() creates a shared, read-only map. Unfortunately there's a quirk in Linux
that prevents the creation of such maps if the file is sealed for write.
(which is a reasonable thing the compositor might want to do).

map_copy_read_only() creates a private, read-only map which doesn't have
this problem.