seat/pointer: simplify ThemedPointer handling
This should simplify the handling of custom cursors in the user
applications by incapsulating all the necessary state in the
ThemedPointer.

The new cursor-icon crate for better compatibility across the
ecosystem, so the icon could be passed as is.

In the future we can automatically handle the `wp_cursor_shape`
under the hood.