WIP Update wayland-rs and calloop
`simple_Window` example here workss.

Incorporates the calloop source that was part of wayland-client, updated
for current wayland-rs and wayland-client. Unless we want that somewhere
else?

Updating these crates is a breaking change, so we might as well update
APIs to use io safe types. Perhaps rustix could be used in some places.