Take `OwnedFd` instead of `RawFd` as argument to `receive_to_fd`
This appears to be the only use of `RawFd` in a public API. With this
change, these functions should be safe to call.

I'm not sure if it would make more sense for these functions to not
close the file descriptor. In which case they could take `BorrowedFd`.
But this matches the current implementation, but with an IO safe type.