Add session-lock protocol
I modelled the API here on how layer shell is handled.

`SessionLock` and `SessionLockSurface` destroy their corresponding
Wayland objects on drop. In the case of `SessionLock`, it is incorrect
to do this before `locked`/`finished` are received, but also incorrect
to not destroy it. For security, the compositor remains locked if any
of these things are done incorrectly.

Ideally `SessionLock` then might be a linear type that you can't drop
but have to pass to an explicit destroy function, but in lieu of such
functionality, it's marked `#[must_use]`.

The example simply creates a lock, creates a lock surface for each
output with the same colorful graphic as `simple_window`, and terminates
after 5 seconds.