Require an explicit unlock call in SessionLock
The entire reason unlock_and_destroy is distinct from destroy is to
avoid accidental unlock operations.  The current SCTK implementation
carefully undoes this work by unlocking on drop, even if that drop is
due to a panic or careless drop of the lock object.

Add an explicit unlock call so that screen lockers can indicate when
they intend to unlock (generally after some kind of authentication has
happened).