Add Clone for dmabuf feedback structs
The `DmabufHandler::dmabuf_feedback` handler provides access to DMA
buffer feedback, however working with this feedback becomes difficult
when it needs to be stored in multiple places due to missing derives.

So this patch adds the automatic `Clone` derive so the existing types in
SCTK can be reused downstream.

The `DmabufFeedback` itself does not implement `Clones` since it relies
on the `Mmap` type, which is not cloneable.