[WIP] Expand the tablet example for pseudo-drawing
I want something *like* this,
partly because it’s a more interesting demo for tablets,
and partly because I want something showing overdrawing.

The implementation is terrible in various ways that I know,
and probably terrible in various ways that I don’t yet know.
But it works… mostly. And it *normally* doesn’t segfault.
I’m just committing it to get *something* out there for now,
not because it’s finished.

Pretty sure I’m going to ditch the mode stuff and put at least the
tablet tool info, state and circle thing into a cursor.
That’s cool.