Tablet info events: new approach, always collect
I’m not *certain* why tablet-v2 uses a burst of events for descriptions.
For frames it makes inherent sense,
as that’s often the way you want to view it,
and for some things close to necessary, if you want it to be meaningful,
but for the static device descriptions it doesn’t make *inherent* sense.

My guesses are:

(1) The information isn’t all available at once.
    I don’t believe this is correct, or meaningfully so.

(2) Wayland protocol wire format limitations.
    I don’t know if it can express a list of strings¹, like paths?

(3) Forwards-compatibility measure.
    Is adding a new event to the burst easier than
    adding a new argument to an existing event?

It would be good to know whether it’s just vagaries of Wayland,
in which case reconstructing the pieces is preferrable,
or whether there is enduring purpose to the fragmentation.

Effects of this change:

  • Ergonomics: significant improvement for direct usage,
    much of a muchness for abstraction layers wanting to use their own struct.

  • Idle memory usage: reduced by 40 bytes per tablet.

  • Runtime: will typically make one more allocation than before.
    Fixable by changing paths from Vec<String> to SmallVec<[String; 1]>,
    in which case runtime should overall be slightly better than before.
    But I’m not sure if we want to expose SmallVec like that.

—⁂—

¹ <https://wayland.freedesktop.org/docs/html/ch04.html#sect-Protocol-Wire-Format:~:text=array>,
  it’s just called “array” and it’s not clear to me what it can contain.