More shuffling tool stuff around, for the better
I think this is pretty close to what people will want now.

I came via passing around &mut Option<State> and such,
but realised it was better to stop blatting *all* the state,
and just shift the proximity_in information into an optional substruct.
This lets things be just methods on State, for much greater consistency,
and the scoping makes it easier to ignore State if you don’t want to use it, too.

I didn’t like the Tools map, it was too inflexible.
So, ditched it in favour of making the user choose the map and struct.
It’s much of a muchness on boilerplate, and simpler, which is a virtue.
Now you see a Tool struct which can have more per-tool fields added.
This is good.

It’s funny how you can come back to earlier designs like this.
This is very similar to what I had a few days ago,
but it was just a *little* different then, in bad ways.
Now I’ve got the best of all worlds I’ve thought of so far.

—⁂—

I’m still mulling over the udata pattern,
which would probably be more consistent with the rest of sctk.
I’m not entirely sold on shifting things into there,
for a couple of initial reasons:

  • I’m not sure if State can be optional quite so well—
    in order to avoid overhead if it’s not being used, it wants
    relaxed coherence, negative bounds, specialisation or similar.

  • If Info remains in the data, it must be in a mutex.
    Mutex guards are nasty to deal with, and compromise API.